Remove setCache call during rendering
The call to setCache to store loading = true is performed during the rendering of the component.
As it is stored in a global reactive variable, it sometimes causes issues of update a component while rendering another one.
https://reactjs.org/link/setstate-in-render

To fix the issue, the setCache is removed and we rely on apollo to deduplicate the requests.